03_rag/requirements.txt
langchain-community
pypdf
chromadb
langchain-groq
python-dotenv
sentence-transformers
langchain-chroma
faiss-cpu
streamlit